FileManager

Description

Module provides file service.

Documentation

rewriteFile :: FilePath -> FilePath -> ([Char] -> [Char]) -> IO ()

The rewriteFile is function which perform reading from file, then converting data with given function and rewrite new data to second file.