MillerRabin

Description

Module provides service for checking if given numbers is prime.

Documentation

isPrime :: Integer -> Bool

The isPrime function is a wrapper for Miller-Rabin

testMillerRabin :: Int -> Integer -> IO Bool

The testMillerRabin tests primality of numer using Miller-Rabin Algoritm

test :: Integral nat => nat -> nat -> [nat] -> nat -> nat -> Bool

The test tests if number is in given sets

witnesses :: (Num a, Ord a, Random a) => Int -> a -> IO [a]

The witnesses returns witnesses for numer which are sufficient to check

powerMod :: Integral nat => nat -> nat -> nat -> nat

The powerMod returns x which satisfy x^n mod m