RSA

Description

Documentation

getMaximumPower :: Int -> Int -> Int

The getMaximumPower function returns the largest p, where 2^p <= number.

getPowerList :: Int -> [Int]

The getPowerList function returns list of powers. Those sum of those 2^power is equal given number.

doModuloTo2 :: Int -> Int -> Int -> Int

The doModuloTo2 function perform operation of exponentation with protection before overflow.

encryptNumber :: Int -> Int -> Int -> Int

The encryptNumber function is encrypting number with key and modulus or decryypting ciphered number with private key an modulus.

encryptText :: Int -> Int -> [Char] -> [Char]

The encryptText function is encryptin plain text with public key and modulus or decrypting cipher with private key and modulus.